feat(classification): upgrade classification module to handle async classification while passing document to LLM
upgrade the classification module to handle the classification asyncronoulsy while passing the document to llm for classification

+++ src/main/java/cs/sonu/TaxDoc/classification/entity/ClassificationResult.java
@@ -1,8 +1,10 @@
 package cs.sonu.TaxDoc.classification.entity;
 
+import java.math.BigDecimal;
+
 import cs.sonu.TaxDoc.document.entity.DocumentType;
 
 public record ClassificationResult(
-        DocumentType documentType,
-        double confidence) {
+                DocumentType documentType,
+                BigDecimal confidence) {
 }
